Switch to friendlier ESLint config
This is based on the ESLint config used by Create React App. It is
much less opinionated than Airbnb, so it mostly provides warnings
about things that are likely bugs, and rarely about code style.

I've also disabled the jsx-a11y plugin, because it also creates a lot
of warnings when you're trying to quickly hack together a program
in a half-day React course.